convert to removeBreakdown action
